added dtp-ps-common
